Implementing option in ecospat.ESM.EnsembleProjection to use only a subset of models
If all bivariate models of a model technqiue failed, or falled below a given performance treshold, the ecospat.ESM.EnsembleProjection function will fail because it tries to average all the models, also the failed one. With this update it is possible to select a subset of techniques for ensembling without re-running all the models. Bug reported by Heidi Mod. Thanks Heidi ;)